// teslo-shop/migrations/snapshots/aa6997e265320471554beada375b7cfb48d1d58659a725c0a2c05aac976beea8/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Category: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly Product: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly inStock: CodecTypes['pg/int4@1']['output'];
      readonly price: CodecTypes['pg/float8@1']['output'];
      readonly sizes: ReadonlyArray<'XS' | 'S' | 'M' | 'L' | 'XL' | 'XXL' | 'XXXL'>;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly tags: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly gender: 'men' | 'women' | 'kid' | 'unisex';
      readonly categoryId: CodecTypes['pg/text@1']['output'];
    };
    readonly ProductImage: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly url: CodecTypes['pg/text@1']['output'];
      readonly productId: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Category: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly Product: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly inStock: CodecTypes['pg/int4@1']['input'];
      readonly price: CodecTypes['pg/float8@1']['input'];
      readonly sizes: ReadonlyArray<'XS' | 'S' | 'M' | 'L' | 'XL' | 'XXL' | 'XXXL'>;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly tags: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly gender: 'men' | 'women' | 'kid' | 'unisex';
      readonly categoryId: CodecTypes['pg/text@1']['input'];
    };
    readonly ProductImage: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly url: CodecTypes['pg/text@1']['input'];
      readonly productId: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly category: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
    readonly product: {
      readonly categoryId: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly gender: 'men' | 'women' | 'kid' | 'unisex';
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly inStock: CodecTypes['pg/int4@1']['output'];
      readonly price: CodecTypes['pg/float8@1']['output'];
      readonly sizes: ReadonlyArray<'XS' | 'S' | 'M' | 'L' | 'XL' | 'XXL' | 'XXXL'>;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly tags: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly title: CodecTypes['pg/text@1']['output'];
    };
    readonly productImage: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly productId: CodecTypes['pg/text@1']['output'];
      readonly url: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly category: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
    readonly product: {
      readonly categoryId: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly gender: 'men' | 'women' | 'kid' | 'unisex';
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly inStock: CodecTypes['pg/int4@1']['input'];
      readonly price: CodecTypes['pg/float8@1']['input'];
      readonly sizes: ReadonlyArray<'XS' | 'S' | 'M' | 'L' | 'XL' | 'XXL' | 'XXXL'>;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly tags: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly title: CodecTypes['pg/text@1']['input'];
    };
    readonly productImage: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly productId: CodecTypes['pg/text@1']['input'];
      readonly url: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;